Refactor NavBar into class component, and move handling fetching user data to it from Posts
