+++ client/src/components/containers/NavBarContainer.js
@@ -11,6 +11,9 @@ const mapDispatchToProps = (dispatch) => ({
   onGetUserFromServer(user) {
     dispatch(actions.getUserFromServer(user));
   },
+  onGetUsersPostVotesFromServer(postVotes) {
+    dispatch(actions.getUsersPostVotesFromServer(postVotes));
+  },
   onUnmountUser() {
     dispatch(actions.unmountUser());
   },
